mm: get mm_bitmap_find_and allocate under control
We had a nifty bug here: If the page_offset is 0, then we interepreted
it as "don't care", which is not always the case.

Better use -1 to signalise "don't care".

Use this chance to get types under control.

+++ include/grinch/mm.h
@@ -15,8 +15,8 @@
 
 #include <grinch/bitmap.h>
 
-/* If from (bit) is not zero, then the area MUST start from start. */
-long mm_bitmap_find_and_allocate(struct bitmap *bitmap, size_t pages, unsigned
-				 int from, unsigned int alignment);
+/* The area MUST start from from, if from != -1 */
+long mm_bitmap_find_and_allocate(struct bitmap *bitmap, unsigned int pages,
+				 long from, unsigned int alignment);
 
 #endif /* _MM_H */
